+++ src/Containers/Home1/Home.js
@@ -1,0 +1,23 @@
+import React from 'react'
+import logo from './logo.svg'
+
+class Home extends React.Component {
+  render () {
+    return (
+      <div className='App'>
+        <div className='App-header'>
+          <img src={logo} className='App-logo' alt='logo' />
+          <h2>Welcome to React</h2>
+        </div>
+        <p className='App-intro'>
+        To get started, edit <code>src/App.js</code> and save to reload.
+      </p>
+        <p>
+        RAHUL WAS HERE
+      </p>
+      </div>
+    )
+  }
+}
+
+export default Home
